chg: moved all routines for tropical geometry to tropical.cc/.h
