+++ Singular/dyn_modules/gfanlib/bbcone.cc
@@ -9,15 +9,10 @@
 #include <coeffs/longrat.h>
 
 #include <Singular/ipid.h>
-#include <Singular/ipid.h>
 #include <Singular/ipshell.h>
 #include <Singular/blackbox.h>
 
-// #include <omalloc/omalloc.h>
-// #include <kernel/intvec.h>
-// #include <kernel/longrat.h>
-// #include <Singular/lists.h>
-// #include <Singular/subexpr.h>
+#include <sstream>
 
 #include <gfanlib/gfanlib.h>
 #include <gfanlib/gfanlib_q.h>
@@ -108,6 +103,14 @@ gfan::ZVector* bigintmatToZVector(const bigintmat &bim)
     n_Delete(&temp,coeffs_BIGINT);
     delete gi;
   }
+  return zv;
+}
+
+gfan::ZVector intStar2ZVector(const int d, const int* i)
+{
+  gfan::ZVector zv(d);
+  for(int j=0; j<d; j++)
+    zv[j]=i[j+1];
   return zv;
 }
 
@@ -1592,56 +1595,6 @@ BOOLEAN containsCone(leftv res, leftv args)
   return TRUE;
 }
 
-gfan::ZVector intStar2ZVector(const int d, const int* i)
-{
-  gfan::ZVector zv(d);
-  for(int j=0; j<d; j++)
-    zv[j]=i[j+1];
-  return zv;
-}
-
-BOOLEAN maximalGroebnerCone(leftv res, leftv args)
-{
-  leftv u = args;
-  if ((u != NULL) && (u->Typ() == IDEAL_CMD))
-  {
-    leftv v = u->next;
-    if (v == NULL)
-    {
-      int n = currRing->N;
-      ideal I = (ideal) u->Data();
-      poly g = NULL;
-      int* leadexpv = (int*) omAlloc((n+1)*sizeof(int));
-      int* tailexpv = (int*) omAlloc((n+1)*sizeof(int));
-      gfan::ZVector leadexpw = gfan::ZVector(n);
-      gfan::ZVector tailexpw = gfan::ZVector(n);
-      gfan::ZMatrix inequalities = gfan::ZMatrix(0,n);
-      for (int i=0; i<IDELEMS(I); i++)
-      {
-        g = (poly) I->m[i]; pGetExpV(g,leadexpv);
-        leadexpw = intStar2ZVector(n, leadexpv);
-        pIter(g);
-        while (g != NULL)
-        {
-          pGetExpV(g,tailexpv);
-          tailexpw = intStar2ZVector(n, tailexpv);
-          inequalities.appendRow(leadexpw-tailexpw);
-          pIter(g);
-        }
-      }
-      gfan::ZCone* gCone = new gfan::ZCone(inequalities,gfan::ZMatrix(0, inequalities.getWidth()));
-      omFreeSize(leadexpv,(n+1)*sizeof(int));
-      omFreeSize(tailexpv,(n+1)*sizeof(int));
-
-      res->rtyp = coneID;
-      res->data = (void*) gCone;
-      return FALSE;
-    }
-  }
-  WerrorS("maximalGroebnerCone: unexpected parameters");
-  return TRUE;
-}
-
 
 lists listOfFacets(const gfan::ZCone &zc)
 {
@@ -1695,163 +1648,12 @@ BOOLEAN listOfFacets(leftv res, leftv args)
 }
 
 
-poly initial(poly p)
-{
-  poly g = p;
-  poly h = p_Head(g,currRing);
-  poly f = h;
-  long d = p_Deg(g,currRing);
-  pIter(g);
-  while ((g != NULL) && (p_Deg(g,currRing) == d))
-  {
-    pNext(h) = p_Head(g,currRing);
-    pIter(h);
-    pIter(g);
-  }
-  return(f);
-}
-
-
-BOOLEAN initial(leftv res, leftv args)
-{
-  leftv u = args;
-  if ((u != NULL) && (u->Typ() == POLY_CMD))
-  {
-    leftv v = u->next;
-    if (v == NULL)
-    {
-      poly p = (poly) u->Data();
-      res->rtyp = POLY_CMD;
-      res->data = (void*) initial(p);
-      return FALSE;
-    }
-  }
-  if ((u != NULL) && (u->Typ() == IDEAL_CMD))
-  {
-    leftv v = u->next;
-    if (v == NULL)
-    {
-      ideal I = (ideal) u->Data();
-      ideal inI = idInit(IDELEMS(I));
-      poly g;
-      for (int i=0; i<IDELEMS(I); i++)
-      {
-        g = (poly) I->m[i];
-        inI->m[i]=initial(g);
-      }
-      res->rtyp = IDEAL_CMD;
-      res->data = (void*) inI;
-      return FALSE;
-    }
-  }
-  WerrorS("initial: unexpected parameters");
-  return TRUE;
-}
-
-
-BOOLEAN homogeneitySpace(leftv res, leftv args)
-{
-  leftv u = args;
-  if ((u != NULL) && (u->Typ() == IDEAL_CMD))
-  {
-    leftv v = u->next;
-    if (v == NULL)
-    {
-      int n = currRing->N;
-      ideal I = (ideal) u->Data();
-      poly g;
-      int* leadexpv = (int*) omAlloc((n+1)*sizeof(int));
-      int* tailexpv = (int*) omAlloc((n+1)*sizeof(int));
-      gfan::ZVector leadexpw = gfan::ZVector(n);
-      gfan::ZVector tailexpw = gfan::ZVector(n);
-      gfan::ZMatrix equations = gfan::ZMatrix(0,n);
-      for (int i=0; i<IDELEMS(I); i++)
-      {
-        g = (poly) I->m[i]; pGetExpV(g,leadexpv);
-        leadexpw = intStar2ZVector(n, leadexpv);
-        pIter(g);
-        while (g != NULL)
-        {
-          pGetExpV(g,tailexpv);
-          tailexpw = intStar2ZVector(n, tailexpv);
-          equations.appendRow(leadexpw-tailexpw);
-          pIter(g);
-        }
-      }
-      gfan::ZCone* gCone = new gfan::ZCone(gfan::ZMatrix(0, equations.getWidth()),equations);
-      omFreeSize(leadexpv,(n+1)*sizeof(int));
-      omFreeSize(tailexpv,(n+1)*sizeof(int));
-
-      res->rtyp = coneID;
-      res->data = (void*) gCone;
-      return FALSE;
-    }
-  }
-  WerrorS("homogeneitySpace: unexpected parameters");
-  return TRUE;
-}
-
-
-BOOLEAN groebnerCone(leftv res, leftv args)
-{
-  leftv u = args;
-  if ((u != NULL) && (u->Typ() == IDEAL_CMD))
-  {
-    leftv v = u->next;
-    if (v == NULL)
-    {
-      int n = currRing->N;
-      ideal I = (ideal) u->Data();
-      poly g = NULL;
-      int* leadexpv = (int*) omAlloc((n+1)*sizeof(int));
-      int* tailexpv = (int*) omAlloc((n+1)*sizeof(int));
-      gfan::ZVector leadexpw = gfan::ZVector(n);
-      gfan::ZVector tailexpw = gfan::ZVector(n);
-      gfan::ZMatrix inequalities = gfan::ZMatrix(0,n);
-      gfan::ZMatrix equations = gfan::ZMatrix(0,n);
-      long d;
-      for (int i=0; i<IDELEMS(I); i++)
-      {
-        g = (poly) I->m[i]; pGetExpV(g,leadexpv);
-        leadexpw = intStar2ZVector(n, leadexpv);
-        pIter(g);
-        d = p_Deg(g,currRing);
-        while ((g != NULL) && (p_Deg(g,currRing) == d))
-        {
-          pGetExpV(g,tailexpv);
-          tailexpw = intStar2ZVector(n, tailexpv);
-          equations.appendRow(leadexpw-tailexpw);
-          pIter(g);
-        }
-
-        if (g != NULL)
-        {
-          while (g != NULL)
-          {
-            pGetExpV(g,tailexpv);
-            tailexpw = intStar2ZVector(n, tailexpv);
-            inequalities.appendRow(leadexpw-tailexpw);
-            pIter(g);
-          }
-        }
-      }
-      gfan::ZCone* gCone = new gfan::ZCone(inequalities,equations);
-      omFreeSize(leadexpv,(n+1)*sizeof(int));
-      omFreeSize(tailexpv,(n+1)*sizeof(int));
-
-      res->rtyp = coneID;
-      res->data = (void*) gCone;
-      return FALSE;
-    }
-  }
-  WerrorS("groebnerCone: unexpected parameters");
-  return TRUE;
-}
-
 /***
  * Given a cone and a point in its boundary,
  * returns the inner normal vector of a facet
  * containing the point.
+ * Unless the point is in the relative interior of the facet
+ * the facet is not unique.
  * In case no facet contains the point,
  * then 0 is returned.
  **/
@@ -1958,10 +1760,6 @@ void bbcone_setup(SModulFunctions* p)
   p->iiAddCproc("","uniquePoint",FALSE,uniquePoint);
   p->iiAddCproc("","listContainsCone",FALSE,containsCone);
   p->iiAddCproc("","listOfFacets",FALSE,listOfFacets);
-  p->iiAddCproc("","maximalGroebnerCone",FALSE,maximalGroebnerCone);
-  p->iiAddCproc("","groebnerCone",FALSE,groebnerCone);
-  p->iiAddCproc("","initial",FALSE,initial);
-  p->iiAddCproc("","homogeneitySpace",FALSE,homogeneitySpace);
   p->iiAddCproc("","facetContaining",FALSE,facetContaining);
   coneID=setBlackboxStuff(b,"cone");
 }
