+++ Singular/dyn_modules/gfanlib/gfanlib.cc
@@ -6,6 +6,7 @@
 #include <bbfan.h>
 #include <bbpolytope.h>
 #include <gitfan.h>
+#include <tropical.h>
 
 #include <Singular/ipid.h>
 #include <Singular/mod_lib.h>
@@ -22,6 +23,7 @@ extern "C" int SI_MOD_INIT(gfanlib)(SModulFunctions* p)
   bbfan_setup(p);
   bbpolytope_setup(p);
   gitfan_setup(p);
+  tropical_setup(p);
   return MAX_TOK;
 }
