+++ Singular/dyn_modules/gfanlib/tropical.h
@@ -1,0 +1,8 @@
+#ifndef TROPICAL_H
+#define TROPICAL_H
+
+#include "Singular/ipid.h"
+
+void tropical_setup(SModulFunctions* p);
+
+#endif
